refactor: 💡 <ChartWidgetFilter> responsibility
