+++ src/components/Widget/types.ts
@@ -8,7 +8,5 @@ export type RenderOptions = {
   isHighlighted: boolean;
   isFadeOut: boolean;
   title?: string;
-  datePickerData?: null | Record<string, any>;
-  hasInterimQuery: boolean;
   onRemoveWidget: () => void;
 };
